+++ src/clj/matcher_combinators/midje.clj
@@ -1,8 +1,11 @@
 (ns matcher-combinators.midje
   (:require [matcher-combinators.core :as core]
+            [matcher-combinators.dispatch :as dispatch]
+            [matcher-combinators.matchers :as matchers]
             [matcher-combinators.model :as model]
             [matcher-combinators.parser]
             [matcher-combinators.result :as result]
+            [matcher-combinators.utils :as utils]
             [midje.data.metaconstant] ; otherwise Metaconstant class cannot be found
             [matcher-combinators.printer :as printer]
             [midje.checking.core :as checking]
@@ -31,6 +34,52 @@
                                (checking/as-data-laden-falsehood
                                 {:notes [(str "Input wasn't a matcher: " matcher)]}))))
 
+(defmacro ^{:doc "Validates that the provided values satisfies the matcher but
+                 uses the provided type->matcher map to redefine the default
+                 matchers used for the specified types.
+
+                 By default when the system sees a `java.lang.Long` it applies
+                 the `equals` matcher to it.  So if, for example, we want to
+                 check that all Longs are greater than whatever Long provided
+                 in the matcher, we would do:
+
+                 `(defn greater-than-matcher [expected-long]
+                    (matcher-combinators.core/->PredMatcher
+                      (fn [actual] (> actual expected-long))))
+
+                 (match-with {java.lang.Long greater-than-matcher})`
+
+                 NOTE: currently doesn't work in midje `provided` expressions"
+            :arglists '([type->default-matcher]
+                        [type->default-matcher matcher])}
+  match-with
+  [& args]
+  (let [arg-count  (count args)
+        ;; Not exactly hygenic or whatever, but that'll do pig, that'll do.
+        ;; Needed because I don't know how to deal with auto-gensym-ing in the
+        ;; context of nested quoting
+        actual-var (gensym 'actual)]
+    (case arg-count
+      1 (let [[type->default-matcher] args
+              matcher-var             (gensym 'mathcer)]
+          `(fn [~matcher-var]
+             (fn [~actual-var]
+               ~(dispatch/match-with-inner
+                 type->default-matcher
+                 `(if (core/matcher? ~matcher-var)
+                    (check-match ~matcher-var ~actual-var)
+                    (checking/as-data-laden-falsehood
+                     {:notes [(str "Input wasn't a matcher: " ~matcher-var)]}))))))
+      2 (let [[type->default-matcher matcher] args]
+          `(fn [~actual-var]
+             ~(dispatch/match-with-inner
+               type->default-matcher
+               `(if (core/matcher? ~matcher)
+                  (check-match ~matcher ~actual-var)
+                  (checking/as-data-laden-falsehood
+                   {:notes [(str "Input wasn't a matcher: " ~matcher)]})))))
+      (throw (ArityException. arg-count "expected 1 or 2 arguments")))))
+
 (defn- parse-throws-args! [args]
   (let [arg-count (count args)]
     (case arg-count
@@ -47,7 +96,7 @@
 
 (checkers.defining/defchecker throws-match
   "Takes in a matcher or a matcher and throwable subclass.
-                              Returns a checker that asserts an exception was raised and the ex-data within it satisfies the matcher"
+  Returns a checker that asserts an exception was raised and the ex-data within it satisfies the matcher"
   [& args]
   (checkers.defining/checker [actual]
                              (if-not (instance? ICapturedThrowable actual)
@@ -86,3 +135,22 @@
          ::result/value  mismatch-val
          ::result/weight 1}))))
 
+(def match-equals
+  "match but using strict `equals` matching behavior for maps, even nested ones."
+  (match-with {clojure.lang.IPersistentMap matchers/equals}))
+
+(def match-roughly
+  "match where all numbers match if they are within the delta of their expected value"
+  (fn [delta matcher]
+    (let [func (fn [expected] (core/->PredMatcher (fn [actual]
+                                                    (utils/roughly? expected actual delta))))]
+      (match-with
+       {java.lang.Integer    func
+        java.lang.Short      func
+        java.lang.Long       func
+        java.lang.Float      func
+        java.lang.Double     func
+        java.math.BigDecimal func
+        java.math.BigInteger func
+        clojure.lang.BigInt  func}
+       matcher))))